added type-hint to eliminate reflection
I noticed the fps for this demo was like
20-30 on my weak laptop, and I was curious
to see why it was so slow.  Turns out it
was the call to near-entity, which was
causing clojure to reflect on the args to
(rectangle! ....) which expect two instances
of gdx Rectangle objects.  I just plugged
in a typehint on entity-rect, and my fps
jumped to a solid 60 once the reflection
call was eliminated.  Easy fix.